Generator update

+++ src/modules/Zikula/ContentModule/Helper/Base/AbstractWorkflowHelper.php
@@ -295,15 +295,17 @@ abstract class AbstractWorkflowHelper
             }
         }
         $result = false;
+        if (!$workflow->can($entity, $actionId)) {
+            return $result;
+        }
     
         try {
-            $workflow->apply($entity, $actionId);
-    
             if ('delete' == $actionId) {
                 $entityManager->remove($entity);
             } else {
                 $entityManager->persist($entity);
             }
+            $workflow->apply($entity, $actionId);
             $entityManager->flush();
     
             $result = true;
